pyproject.toml
[project]
name = "llm2sql"
version = "0.1.0"
description = "Benchmark SQL generation from LLMs against ground-truth queries"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0",
    "modal>=0.67",
    "streamlit>=1.30",
    "pandas>=2.0",
    "plotly>=5.0",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
]
